Fix lock acquiring behaviour in `HTTPClient.request`.
Previously, this method would acquire the lock for all
requests, including requests which didn't need to use
the cache. This would cause long-standing requests to
block the HTTP client from making any further requests
until that request finishes. This commit fixes this
behaviour to only acquire the lock when caching.