Rewrite `human_delta`.
This commit refactors `human_delta` to use `relativedelta` instead
of `divmod`. As part of this change, this will replace the `delta`
argument with the `datetime1` and `datetime2` arguments, both of
which take ``datetime`` objects. This improves the accuracy of
this method since two `datetime` objects provide more information
than a singular `timedelta` object. Furthermore, calculations of
units beyond weeks are now no longer done using estimates since
all of this is now handled through `relativedelta`.

This also fixes a documentation nit where a non-existant keyword
argument ``indicate_tense`` was shown in the examples.