Fix `actual_max_size` not dynamically updating.
This essentially changes `actual_max_size` to be a read-only
property rather than an attribute. The reason this is currently
not prompting a breaking change is that I never intended for
anyone to actually modify the value of `actual_max_size` and
that I have no reason to believe that there is any actual usecase
for doing so without just modifying `prefix`, `suffix` or even
`max_size` itself.