Fully refactor `tucker` and `soyjak`.
This commit essentially supercedes any changes made in
the following commits:

* b25e9c1bc25b7aa1a19a87d5e616be6a6f31abd0
* 92700bfeb483be091e2144c4d7894e3cd85a54eb
* 08813bc74574c76ac56d87b9107dfa18bae9d311

This provides sizeable speedups (at least, on my end
during testing) for both of the above commands. In
addition, this also fixes a bug in `soyjak` where the
resulting image would keep the transparency of the
given image if it had a 16:9 aspect ratio.