Fix `ytinfo` throwing `ValueError`.
It seems that, for whatever reason, YouTube decided to
include fractions of seconds only in channels created
after some date in 2019(?). All other channels created
prior to this date do not include fractions of seconds
(potentially because they were removed).

The initial commit for v3.0 did attempt to fix what
I thought was YouTube removing fractions of a second
in creation dates for all channels without documenting
it. I'm not entirely sure when this change to the API
was implemented, but I didn't get any heads-up for this
nor was it actually documented anywhere.