Use uvloop if available.
This adds the optional necessary setup to __main__ for
uvloop.

Note that I will not list it under `# Dependencies` in
README or add it to `requirements.txt` since uvloop may
have slight stability issues and/or race conditions. It
also isn't compatible with Windows.

All in all, if you use uvloop with this, YMMV.