Remove useless `per_page` kwarg in `PaginatorSource`.
This was an artifact of an earlier experimental
implementation of this class during the developement
of 3.0 that was ultimately scrapped and rewritten.
Unfortunately, this was forgotten and left in during
the rewrite and thus, made it past several versions.
This kwarg ultimately didn't have a use as its
functionality was never implemented, and setting it
would eventually throw an exception. This commit
shouldn't be a breaking change as I have no reason
to believe that anyone was actively using this.