Comment out `miku`.
The API this is powered by is currently on hiatus
until v3 is completed, in an effort to respect
content creators and their work, since the API
provided these images uncredited. This will be
uncommented when the API comes back, and I'll
make sure to include the credits the API provides
in the resulting embed.