Comply with recent discord.py changes.
This is the first of several commits that will gradually
bring Sleepy up to date with the recent changes to d.py.
I don't intend for this shift to be complete until d.py
reaches its stable release of v2.0.0. This also brings
in some much needed improvements and rewrites to some
old components, as well as a few bugfixes.

The complete list of changes, including those breaking,
will be documented at some point closer to release.