Use jishaku's `WrappedPaginator` instead.
This replaces sleepy's legacy `WrappedPaginator` as everything the
old paginator addressed as been implemented in `jishaku`'s version.
Furthermore, there's also a performance increase.

The only breaking changes are a change in the `__init__` parameter
names and the removal of the `actual_max_size` property, which, in
hindsight, wasn't really helpful anywhere outside of the paginator
itself.