Fix colouring & performance issues in `nichijou`.
Also removes a useless kwarg being passed when saving the GIF.

This occurred due to a change in GIF loading in Pillow 9.x where
subsequent frames are converted to RGB or RGBA. In the interest of
performance, the only sensible way to fix this is to set the GIF
loading strategy to `RGB_AFTER_DIFFERENT_PALETTE_ONLY`, however,
the option to do so wasn't introduced until `9.1.x`, meaning that
Pillow 9.0.x must be skipped. If using Pillow 9.0.x is necessary,
then line 244 in `backend.py` can be changed to the following:

`frames.append(frame.convert("RGBA"))`.