Remove NSFW cog.
Simply put, this cog is a mess that I do not wish to maintain anymore.
Considering the various types of utterly depraved, borderline illegal,
and morally reprehensible shit gets posted under this banner, I don't
want this to bring any potential drama or issues in the future.

TL;DR That stuff is degenerate and I do not wish to continue touching
it with a thirty-nine-and-a-half-foot pole. It's not worth the mental
tax required for upkeep.

I blame my past self for making the mistake of bringing this cog to
fruition.