Rewrite and simplify the extension loading system.
This effectively overhauls the extension loader to consider only
the `extensions_list` configuration variable while also bringing
its syntax in parity with `jishaku`'s extensions commands syntax.
This effectively simplifies the extension loading process as the
previous system had several redundancies and was overly complex
for no good reason.

In essence, this changes `extensions_list` to take a list of dot
separated extension names and adds functionality for `braceexpand`
syntax. Additionally, this also enforces '*' as a special wildcard
character for extension names. As such, extensions cannot be named
'*'. The changes are better explained in the documentation for the
config value in `config_template.yaml`.

The overhaul also resulted in the following removals as these are
artifacts of the old system, which are now obsolete:

- Removed the `extensions_directory` config value.
    - Removed `Sleepy.extensions_directory`
    - Removed '$' as a special character for extension names.
- Removed the `enable_autoload` config value.
- Removed `Sleepy.get_all_extensions()`