Fix image commands to use new API.
Any command that took a single `image` argument has been rewritten to
take two new arguments instead: `attachment` and `url`, both of which
are optional. If neither are given, then the author's display avatar
is used instead for the command. Since the commands will eventually be
reorganized and converted to hybrid commands, this design is not final
and further changes may be made in the future.

Also, involves a slight change to `trapcard`: Both text arguments are
now command flags due to the changes, and various docs revisions.

This should also fix linting.