Adjust image converters API to allow for more ergonomic command design.
This reintroduces the attachment conversion functionality into
`ImageResourceConverter`, which is mostly a reversal from the
changes introduced in 892a5bba9d60bcaad3e11d5ad493bb7c9054ec54.

The reasoning for this change is that the API introduced in the
aforementioned commit often lead to confusing syntax in message
and hybrid commands, as observed during further testing. As such,
the API has been adjusted to allow for separate message and app
commands, which was much more ergonomic compared to the proposed
command design in that commit. As such, `ImageAttachment` still
exists as a standalone transformer only.

Most changes applied in 892a5bba9d60bcaad3e11d5ad493bb7c9054ec54
still remain.